Lists: * Rename GroupSort to ObjectList, move to Lists module - apply the same transformation to GroupSortTest
