+++ lists/src/main/kotlin/mathtools/lists/objects/ObjectList.kt
@@ -1,14 +1,15 @@
-package mathtools.systems.groups
+package mathtools.lists.objects
 
-/** Sorting related Group operations */
-object GroupSort {
+/** Sorting related Group operations
+ * Developed by DK96-OS : 2022 */
+object ObjectList {
 
     /** Remove and return the items in the list that match the selector.
       *  @param list A mutable list which will be modified if any items match
       *  @param presorted If the list is sorted according to selector, ie. the matches are all in one consecutive index range in the list.
       *  @param selector The condition to test for determining a match
       *  @return A new MutableList containing matching items or null if no matches were found */
-    inline fun <T> extractGroup(
+    inline fun <T> extractFrom(
         list:MutableList<T>, presorted:Boolean=false, selector: (T) -> Boolean
     ) : MutableList<T>? {
         var i = list.indexOfFirst { selector(it) }
